requirements.txt
langchain
langchain-community
python-dotenv
ollama
faiss-cpu
